+++ app/api/index.js
@@ -1,11 +1,9 @@
-import { loadWallets } from './wallet/load-wallets';
-import { createPersonalWallet } from './wallet/create-wallet';
-import { loadWalletTransactions } from './wallet/load-wallet-transactions';
-import { sendMoney } from './wallet/send-money';
+import accounts from './accounts';
+import wallets from './wallets';
+import data from './data';
 
 export default {
-  loadWallets,
-  createPersonalWallet,
-  loadWalletTransactions,
-  sendMoney
+  accounts,
+  wallets,
+  data
 };