+++ app/api/wallet/load-wallets.js
@@ -1,6 +1,0 @@
-// @flow
-import wallets from '../data/wallets.json';
-
-export const loadWallets = () => new Promise((resolve) => {
-  resolve(wallets);
-});