+++ app/api/wallets/load-wallets.js
@@ -1,0 +1,8 @@
+// @flow
+import data from '../data';
+
+export const loadWallets = () => new Promise((resolve) => {
+  setTimeout(() => {
+    resolve(data.wallets);
+  }, 1000);
+});