+++ app/components/widgets/LoadingSpinner.js
@@ -1,0 +1,12 @@
+// @flow
+import React, { Component } from 'react';
+import styles from './LoadingSpinner.scss';
+
+export default class LoadingSpinner extends Component {
+
+  root: HTMLElement;
+
+  render() {
+    return <div className={styles.component} ref={(div) => { this.root = div; }} />;
+  }
+}