+++ app/controllers/AppController.js
@@ -1,23 +1,40 @@
 // @flow
 import { action } from 'mobx';
 import type { appState } from '../state/index';
+import AccountController from './AccountController';
 import WalletsController from './WalletsController';
 import SidebarController from './SidebarController';
 
 export default class AppController {
 
   state: appState;
+  account: AccountController;
   wallets: WalletsController;
   sidebar: SidebarController;
+  initializedCallback = () => {};
 
   constructor(state: appState) {
     this.state = state;
+    this.account = new AccountController(state);
     this.wallets = new WalletsController(state);
     this.sidebar = new SidebarController(state);
+
+    this.account.loadAccount();
+    this.wallets.loadWallets();
   }
 
-  @action setAppRouter(router: Object) {
+  onInitialized(callback: () => {}) {
+    this.initializedCallback = callback;
+  }
+
+  @action initialize(router: Object) {
     this.state.router = router;
+    this.initializedCallback();
+  }
+
+  navigateTo(route: string) {
+    if (!this.state.router) throw new Error('Cannot to navigate because router is not yet set.');
+    this.state.router.transitionTo(route);
   }
 
 }