+++ app/controllers/WalletsController.js
@@ -4,6 +4,7 @@ import type { appState } from '../state/index';
 import Wallet from '../domain/Wallet';
 import WalletTransaction from '../domain/WalletTransaction';
 import api from '../api';
+import { INITIAL_WALLET_SEARCH_LIMIT } from '../state/active-wallet';
 
 export default class WalletsController {
 
@@ -14,13 +15,13 @@ export default class WalletsController {
   }
 
   @action async loadWallets() {
-    const { activeWallet, account } = this.state;
+    const { activeWallet, user } = this.state;
     activeWallet.isLoading = true;
     try {
-      const wallets = await api.loadWallets();
+      const wallets = await api.wallets.loadWallets();
       for (const wallet of wallets) {
         const newWallet = new Wallet(wallet);
-        account.addWallet(newWallet);
+        user.addWallet(newWallet);
         if (newWallet.lastUsed) this.setActiveWallet(newWallet);
       }
       activeWallet.isLoading = false;
@@ -29,33 +30,37 @@ export default class WalletsController {
     }
   }
 
-  @action async loadActiveWalletTransactions() {
+  @action async loadActiveWalletTransactions(initialLoading: boolean) {
     const { activeWallet } = this.state;
     const { wallet } = activeWallet;
     if (!wallet === null) throw new Error('No active wallet');
-    try {
-      const transactions = await api.loadWalletTransactions({
-        address: wallet.address
-      });
-      wallet.transactions.clear();
-      for (const transaction of transactions) {
-        wallet.addTransaction(new WalletTransaction(transaction));
-      }
-    } catch (error) {
-      activeWallet.errorLoadingTransactions = error;
-      // TODO: handling errors from backend and i18n
+    activeWallet.isLoadingTransactions = true;
+    if (initialLoading) activeWallet.hasAnyTransactions = false;
+    const result = await api.wallets.loadWalletTransactions({
+      address: wallet.address,
+      searchTerm: activeWallet.transactionsSearchTerm,
+      limit: activeWallet.transactionsSearchLimit
+    });
+    wallet.transactions.clear();
+    for (const transaction of result.transactions) {
+      wallet.addTransaction(new WalletTransaction(transaction));
     }
+    activeWallet.totalAvailableTransactions = result.total;
+    activeWallet.isLoadingTransactions = false;
+    if (initialLoading) activeWallet.hasAnyTransactions = result.total > 0;
   }
 
   @action setActiveWallet(walletId: string|Wallet) {
     let wallet = walletId;
     if (_.isString(walletId)) {
-      wallet = _.find(this.state.account.wallets, { address: wallet });
+      wallet = _.find(this.state.user.wallets, { address: wallet });
     }
     const activeWallet = this.state.activeWallet;
     if (wallet === activeWallet.wallet) return;
     activeWallet.wallet = wallet;
-    this.loadActiveWalletTransactions();
+    activeWallet.transactionsSearchLimit = INITIAL_WALLET_SEARCH_LIMIT;
+    activeWallet.transactionsSearchTerm = '';
+    this.loadActiveWalletTransactions(true);
     if (this.state.router) this.state.router.transitionTo(`/wallet/${wallet.address}/home`);
   }
 
@@ -67,7 +72,7 @@ export default class WalletsController {
     const { activeWallet } = this.state;
     const { wallet } = activeWallet;
     try {
-      const transaction = await api.sendMoney({
+      const transaction = await api.wallets.sendMoney({
         ...transactionDetails,
         amount: parseFloat(transactionDetails.amount),
         sender: wallet.address,
@@ -82,9 +87,9 @@ export default class WalletsController {
 
   @action async createPersonalWallet(newWalletData: { name: string, currency: string}) {
     try {
-      const createdWallet = await api.createPersonalWallet(newWalletData);
+      const createdWallet = await api.wallets.createPersonalWallet(newWalletData);
       const newWallet = new Wallet(createdWallet);
-      this.state.account.addWallet(newWallet);
+      this.state.user.addWallet(newWallet);
       this.setActiveWallet(newWallet);
       // TODO: Navigate to newly created wallet
     } catch (error) {
@@ -92,4 +97,15 @@ export default class WalletsController {
     }
   }
 
+  @action filterTransactions(searchTerm) {
+    const { activeWallet } = this.state;
+    activeWallet.transactionsSearchTerm = searchTerm;
+    activeWallet.wallet.transactions.clear();
+    this.loadActiveWalletTransactions();
+  }
+
+  @action loadMoreTransactions() {
+    this.state.activeWallet.transactionsSearchLimit += 10;
+    this.loadActiveWalletTransactions();
+  }
 }