+++ app/state/active-wallet.js
@@ -1,18 +1,33 @@
 // @flow
+import type { appState } from './index';
 import Wallet from '../domain/Wallet';
 
+export const INITIAL_WALLET_SEARCH_LIMIT = 10;
+
 export type activeWalletState = {
+  state: appState,
   wallet: ?Wallet,
-  isLoading: bool,
+  isLoadingTransactions: boolean,
+  transactionsSearchTerm: string,
+  transactionsSearchLimit: number,
+  totalAvailableTransactions: number,
+  isLoading: boolean,
+  hasAnyTransactions: boolean,
   errorLoading: ?string,
   errorCreating: ?string,
   errorLoadingTransactions: ?string,
   errorSendingMoney: ?string
 };
 
-export default (): activeWalletState => ({
+export default (state: appState): activeWalletState => ({
+  state,
   wallet: null,
+  isLoadingTransactions: false,
+  transactionsSearchTerm: '',
+  transactionsSearchLimit: INITIAL_WALLET_SEARCH_LIMIT,
+  totalAvailableTransactions: 0,
   isLoading: false,
+  hasAnyTransactions: false,
   errorLoading: null,
   errorCreating: null,
   errorLoadingTransactions: null,