+++ features/support/electron.js
@@ -2,7 +2,7 @@ import { Application } from 'spectron';
 import electronPath from 'electron';
 
 export default function () {
-  this.Before(async function() {
+  this.Before({ timeout: 20 * 1000 }, async function() {
     this.app = new Application({
       path: electronPath,
       args: ['./electron/main.testing'],
@@ -15,6 +15,12 @@ export default function () {
     await this.app.start();
     this.client = this.app.client;
     this.browserWindow = this.app.browserWindow;
+    await this.client.waitUntilWindowLoaded();
+    await this.client.executeAsync(function(done) {
+      daedalus.environment.current = 'test';
+      daedalus.api.data.reset();
+      daedalus.controller.onInitialized(() => done());
+    });
   });
   this.After(function() {
     return this.app.stop();