+++ stories/index.js
@@ -2,3 +2,4 @@ import '../app/styles/index.global.scss';
 import './Sidebar.stories';
 import './SidebarCategory.stories';
 import './SidebarWalletsMenu.stories';
+import './ProfileSettings.stories';