+++ CHANGELOG.md
@@ -1,6 +1,32 @@
 Changelog
 =========
 
+## 0.5.0
+
+![Main Wallet screenshot in 0.5.0](screenshots/2016-11-24 release 0.5.0 main-wallet.png)
+![Main Wallet screenshot in 0.5.0](screenshots/2016-11-24 release 0.5.0 profile-settings.png)
+
+### Features
+
+- Infinite loading for transaction list, testable with “Main wallet”
+- Simple transaction search by title
+- Adding new personal wallet via sidebar menu
+- Profile settings screen
+- “No transactions” message for wallets with no transactions
+- “No transactions found” message when transaction search returns no results
+- Loading indicator on application startup
+
+### Fixes
+
+- Improved sidebar UX when there is no sub-menu
+- Transaction icons updated to design specs
+- Improved transaction list page styling
+- Active input field styled to design specs
+
+### Chores
+
+- Fixed and improved first acceptance test, made api data configurable for future test cases
+
 ## 0.4.0
 
 ![Screenshot from Release 0.4.0](screenshots/2016-11-16 release 0.4.0.png)